refactor: removido arquivos desnecessários
