+++ app/page.js
@@ -2,7 +2,7 @@ import Nav from "@/components/menu/Nav";
 import About from "./home/about/About";
 import { Contact } from "./home/contact/Contact";
 import Home from "./home/Home";
-import Projects from "./home/projects/Projects";
+
 import Stacks from "./home/stacks/Stacks";
 import Avatar from "./home/avatar/Avatar";
 
@@ -17,7 +17,6 @@ export default function PageIndex() {
           <div className="col-two">
             <Avatar />
             <About />
-            <Projects />
             <Stacks />
             <Contact />
           </div>
