--- components/CardLink/style.module.css
@@ -1,43 +1,0 @@
-.card {
-  width: 100%;
-  min-height: 100px;
-  display: grid;
-  grid-template-columns: 150px 1fr;
-  gap: 2rem;
-  padding: 2rem;
-  border-radius: var(--radius);
-  margin-bottom: 2rem;
-}
-.cardBody {
-  color: var(--color-text-primary);
-}
-.card h1 {
-  color: var(--color-text-primary);
-  margin: 0 0 0.5rem 0;
-}
-
-.card:hover {
-  background-color: #000;
-}
-.cardImg {
-  width: 250px;
-  overflow: hidden;
-}
-.cardImg img {
-  width: 150px;
-  object-fit: contain;
-  border-radius: var(--radius);
-}
-
-.tecnolgies {
-  display: flex;
-  gap: 12px;
-  margin: 10px 0;
-  flex-wrap: wrap;
-}
-
-@media (max-width: 890px) {
-  .card {
-    grid-template-columns: 1fr;
-  }
-}
